requirements.txt
# Core web framework and runtime dependencies.
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
packaging==25.0
python-dotenv==1.1.1
Werkzeug==3.1.3
